requirements.txt
asgiref==3.8.1
Django==4.2.15
django-cors-headers==4.4.0
django-mongoengine==0.5.6
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.15.2
dnspython==2.6.1
mongoengine==0.28.2
pymongo==4.8.0
sqlparse==0.5.1
tzdata==2024.1
